Changed package structure - api classes moved to a separate bundle
