+++ bundles/org.platformio.eclipse.ide.home/META-INF/MANIFEST.MF
@@ -6,3 +6,7 @@ Bundle-Vendor: %Bundle-Vendor
 Bundle-Name: %Bundle-Name
 Automatic-Module-Name: org.platformio.eclipse.ide.home
 Bundle-RequiredExecutionEnvironment: JavaSE-11
+Require-Bundle: org.eclipse.osgi;bundle-version="0.0.0"
+Export-Package: org.platformio.eclipse.ide.home.api,
+ org.platformio.eclipse.ide.home.core,
+ org.platformio.eclipse.ide.home.python
