+++ bundles/org.platformio.eclipse.ide.home/src/org/platformio/eclipse/ide/home/Activator.java
@@ -1,0 +1,18 @@
+package org.platformio.eclipse.ide.home;
+
+import org.osgi.framework.BundleActivator;
+import org.osgi.framework.BundleContext;
+
+public class Activator implements BundleActivator {
+
+	@Override
+	public void start(BundleContext context) throws Exception {
+
+	}
+
+	@Override
+	public void stop(BundleContext context) throws Exception {
+
+	}
+
+}
