+++ bundles/org.platformio.eclipse.ide.installer/META-INF/MANIFEST.MF
@@ -10,7 +10,8 @@ Require-Bundle: org.apache.httpcomponents.httpcore;bundle-version="0.0.0",
  org.eclipse.core.runtime;bundle-version="0.0.0",
  org.apache.httpcomponents.httpclient;bundle-version="0.0.0",
  org.apache.commons.compress;bundle-version="0.0.0",
- com.google.gson;bundle-version="0.0.0"
+ com.google.gson;bundle-version="0.0.0",
+ org.platformio.eclipse.ide.home
 Export-Package: org.platformio.eclipse.ide.installer,
  org.platformio.eclipse.ide.installer.api,
  org.platformio.eclipse.ide.installer.python
